[winui/uno] improve sourcegen
