+++ src/skiasharp/_Shared.WinUI/_Shared.WinUI.projitems
@@ -10,7 +10,10 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)CartesianChart.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)ChartControl.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)LiveChartsGeneratedCode\CartesianChart.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)LiveChartsGeneratedCode\ChartControl.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)LiveChartsGeneratedCode\PieChart.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)LiveChartsGeneratedCode\PolarChart.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MotionCanvas.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PieChart.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PolarChart.cs" />
